requirements.txt
Django~=3.2.6
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
dj-database-url==0.5.0
django-environ~=0.4.5
dj-static==0.0.6
python-decouple==3.3
